// src/app/models/personModel.ts
export class personModel{
    id?: number | undefined;
    id_type?: string | undefined;
    id_number?: number | undefined;
    name?: string | undefined;
    lastname?: string | undefined;
    cellphone?: number | undefined;
    email?: string | undefined;
    level?: number | undefined;
    username?: string | undefined;
    password?: string | undefined;
    active?: boolean | undefined;
    status?: string | undefined;
}